Creates install_H265Repack.sh
New installer file will download current tagged version of this repository
and then installs H265Repack to user ~/.bin as well as included man page.
If ~/.bin doesn't exist it will create it. If it isn't in the $PATH it will
add it.
The installer also checks for homebrew and ffmpeg and installs them if they
aren't present.
All created temp files are removed and the script has full error handling
to troubleshoot if installation fails.